packed styles and markers to different files
