+++ App.js
@@ -2,10 +2,10 @@ import React, { useState, useEffect } from "react";
 import MapView, { Marker, UrlTile, Overlay } from "react-native-maps";
 import { StyleSheet, Text, View, Dimensions } from "react-native";
 import * as Location from "expo-location";
-
-import Hideandshowcomponent from './shared/Hideandshowcomponent'
+import { globalStyles } from "./styles/global";
+import Hideandshowcomponent from "./shared/Hideandshowcomponent";
 import Tourfooter from "./shared/Tourfooter";
-
+import Markers from "./shared/markers";
 
 export default function App() {
   const [location, setLocation] = useState(null);
@@ -18,27 +18,6 @@ export default function App() {
     setShow(false)
   }*/
 
-  useEffect(() => {
-    (async () => {
-      let { status } = await Location.requestPermissionsAsync();
-      if (status !== "granted") {
-        setErrorMsg("Permission to access location was denied");
-      }
-
-      let location = await Location.getCurrentPositionAsync({});
-      setLocation(location);
-      setLatitude(location.coords.latitude);
-      setLongitude(location.coords.longitude)
-    })();
-  }, []);
-
-  let text = "Waiting..";
-  if (errorMsg) {
-    text = errorMsg;
-  } else if (location) {
-    text = JSON.stringify(location);
-  }
-
   // Map image - can replace urlTile within MapView
   // <Overlay
   //  image={'https://calvin.edu/dotAsset/184d0710-a659-4ef6-bc18-d0ac7d9cd057/'}
@@ -46,96 +25,30 @@ export default function App() {
   // />
 
   return (
-    <View style={styles.container}>
-      <MapView
-        // region={this.state.region}
-        // onRegionChange={this.onRegionChange}
-        style={styles.mapStyle}
-        showsUerLocation={true}
-        followsUserLocation={true}
-        region={{
-          latitude: latitude,
-          longitude: longitude,
-          latitudeDelta: 0.0032,
-          longitudeDelta: 0.0031,
-        }}
-      >
-        <UrlTile
-          urlTemplate={'http://c.tile.openstreetmap.org/{z}/{x}/{y}.png'}
-        />
-        
-
-        <Marker
-          coordinate={{ latitude: 42.930548, longitude: -85.58581 }}
-          title="Stop 1"
-        ></Marker>
-        <Marker
-
-          coordinate={{ latitude: latitude, longitude: longitude }}
-          pinColor={'blue'}
-          title="You are here"
-
-          coordinate={{ latitude: 42.92965, longitude: -85.58762 }}
-          title="Stop 2"
-        ></Marker>
-        <Marker
-          coordinate={{ latitude: 42.92930, longitude: -85.58845 }}
-          title="Stop 3"
-        ></Marker>
-        <Marker
-          coordinate={{ latitude: 42.93095, longitude: -85.58926 }}
-          title="Stop 4"
-        ></Marker>
-        <Marker
-          coordinate={{ latitude: 42.93301, longitude: -85.58917 }}
-          title="Stop 5"
-        ></Marker>
-        <Marker
-          coordinate={{ latitude: 42.93330, longitude: -85.58635 }}
-          title="Stop 6"
-        ></Marker>
-        <Marker
-          coordinate={{ latitude: 42.93125, longitude: -85.58701 }}
-          title="Stop 7"
-        ></Marker>
-        <Marker
-          coordinate={{latitude: latitude, longitude: longitude }}
-          pinColor = {'blue'}  
-          title="You are here"        
-        >
-          {/* <View style={{
-                    backgroundColor:'white',
-                    borderWidth: 1,
-                }}>
-                    <Text>You are here</Text>
-                </View> */}
-
-        </Marker>
-      </MapView>
-
-      <Hideandshowcomponent/>
-      <Tourfooter/>
+    <View style={globalStyles.container}>
+      <Markers />
+      <Hideandshowcomponent />
+      <Tourfooter />
     </View>
   );
 }
 
 // {show ? <Hideandshowcomponent changeShow={changeShow}/> : null}
 
-const styles = StyleSheet.create({
-  container: {
-    zIndex: 5,
-    flex: 1,
-    backgroundColor: "#fff",
-    alignItems: "center",
-    justifyContent: "center",
-  },
-  mapStyle: {
-    zIndex: 10,
-    width: Dimensions.get("window").width,
-    height: Dimensions.get("window").height,
-  },
-  coordinates: {
-    fontSize: 10,
-  },
-});
-
+// const styles = StyleSheet.create({
+//   container: {
+//     zIndex: 5,
+//     flex: 1,
+//     backgroundColor: "#fff",
+//     alignItems: "center",
+//     justifyContent: "center",
+//   },
+//   mapStyle: {
+//     zIndex: 10,
+//     width: Dimensions.get("window").width,
+//     height: Dimensions.get("window").height,
+//   },
+//   coordinates: {
+//     fontSize: 10,
+//   },
+// });
