+++ styles/global.js
@@ -1,0 +1,20 @@
+import { StyleSheet } from "react-native";
+import { Dimensions } from "react-native";
+
+export const globalStyles = StyleSheet.create({
+  container: {
+    zIndex: 5,
+    flex: 1,
+    backgroundColor: "#fff",
+    alignItems: "center",
+    justifyContent: "center",
+  },
+  mapStyle: {
+    zIndex: 10,
+    width: Dimensions.get("window").width,
+    height: Dimensions.get("window").height,
+  },
+  coordinates: {
+    fontSize: 10,
+  },
+});
